유니온, 같은 링크 처리 최적화
충돌 정확도를 활용한 타일 충돌
음악

+++ KCore/KBitMap_Animator.cpp
@@ -75,7 +75,7 @@ bool KBitMap_Animator::init()
 {
 	KRenderer::init();
 	
-	if (iAniIdx >= CurAniIter->second.size())
+	if (iAniIdx >= (int)CurAniIter->second.size())
 	{
 		iAniIdx = 0;
 		fAniTime = .0f;
@@ -94,7 +94,7 @@ void KBitMap_Animator::update()
 		fAniTime = .0f;
 	}
 
-	if (CurAniIter->second.size() <= iAniIdx)
+	if ((int)CurAniIter->second.size() <= iAniIdx)
 	{
 		iAniIdx = 0;
 	}
@@ -106,7 +106,7 @@ void KBitMap_Animator::render()
 		return;
 	}
 
-	if (CurAniIter->second.size() <= iAniIdx)
+	if ((int)CurAniIter->second.size() <= iAniIdx)
 	{
 		iAniIdx = 0;
 	}
